Remove unnecessary path.isfile assertion.
Under certain circumstances (in particular multiprocess builds, when testing runs in parallel with other build tasks like documentation builds), files found when scanning may not exist by the time they are passed to the `_collectfile()` routine.  It is not an error for this to occur, so the assertion is excessive and leads to intermittent failures.  Instead, merely returning an empty list is safe and avoids the race condition.